+++ backend/demo/src/test/java/com/sesac/solbid/service/OAuth2ServiceTest.java
@@ -2,11 +2,11 @@ package com.sesac.solbid.service;
 
 import com.sesac.solbid.domain.User;
 import com.sesac.solbid.domain.enums.UserType;
-import com.sesac.solbid.dto.OAuth2Dto;
 import com.sesac.solbid.dto.UserDto;
 import com.sesac.solbid.exception.ErrorCode;
 import com.sesac.solbid.exception.OAuth2Exception;
 import com.sesac.solbid.util.JwtUtil;
+import com.sesac.solbid.dto.auth.response.AuthUrlResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -112,7 +112,7 @@ class OAuth2ServiceTest {
         when(urlGenerator.generateAuthUrl(provider, expectedState)).thenReturn(expectedAuthUrl);
 
         // When
-        OAuth2Dto.AuthUrlResponse response = oAuth2Service.generateAuthUrl(provider);
+        AuthUrlResponse response = oAuth2Service.generateAuthUrl(provider);
 
         // Then
         assertThat(response).isNotNull();
@@ -136,7 +136,7 @@ class OAuth2ServiceTest {
         when(urlGenerator.generateAuthUrl(provider, expectedState)).thenReturn(expectedAuthUrl);
 
         // When
-        OAuth2Dto.AuthUrlResponse response = oAuth2Service.generateAuthUrl(provider);
+        AuthUrlResponse response = oAuth2Service.generateAuthUrl(provider);
 
         // Then
         assertThat(response).isNotNull();
